backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
anthropic==0.40.0
chromadb==0.5.23
sentence-transformers==3.3.1
PyMuPDF==1.25.1
pydantic==2.10.3
